requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
basemap==1.4.1
basemap-data==1.3.2
basemap-data-hires==1.3.2
bcrypt==4.1.3
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
bleach==6.1.0
blinker==1.7.0
bokeh==3.4.1
branca==0.7.2
build==1.2.1
cachetools==5.3.3
caer==2.0.8
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
cmake==3.29.3
colorama==0.4.6
colorcet==3.1.0
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.1
configparser==7.0.0
constantly==23.10.4
contourpy==1.2.0
coverage==6.5.0
coveralls==3.3.1
cryptography==42.0.5
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.5.14
datasets==2.18.0
dateutils==0.6.12
debugpy==1.8.1
decorator==5.1.1
deepface==0.0.91
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dnspython==2.4.2
docopt==0.6.2
docx2txt==0.8
duckdb==0.10.1
email-validator==2.1.0.post1
exceptiongroup==1.2.0
executing==2.0.1
face-alignment==1.4.1
faiss-cpu==1.8.0
fastapi==0.104.1
filelock==3.13.1
fiona==1.9.6
fire==0.6.0
Flask==3.0.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.45.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gast==0.5.4
gdown==5.2.0
genson==1.2.2
geopandas==0.14.4
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.30.0
google-pasta==0.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.0
gunicorn==22.0.0
gw_dsl_parser==0.1.47
h11==0.14.0
h5py==3.11.0
holoviews==1.18.3
htmltools==0.5.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hvplot==0.9.2
hydra-core==1.3.2
hyperlink==21.0.0
hypothesis==6.100.1
idna==3.4
imageio==2.34.1
imbalanced-learn==0.12.0
imblearn==0.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
imutils==0.5.4
incremental==22.10.0
iniconfig==2.0.0
install-jdk==1.1.0
intel-openmp==2021.4.0
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.2
itsdangerous==2.1.2
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_bokeh==4.0.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kanaries_track==0.0.4
keras==3.3.3
Keras-Applications==1.0.8
keras-facenet==0.3.2
keras-vggface==0.6
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-huggingface==0.0.3
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langsmith==0.1.79
lazy_loader==0.4
libclang==18.1.1
linkify-it-py==2.0.3
llvmlite==0.42.0
loguru==0.7.2
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mediapipe==0.10.14
mizani==0.11.1
mkl==2021.4.0
ml-dtypes==0.3.2
mmh3==4.1.0
mock==5.1.0
monotonic==1.6
mplcyberpunk==0.7.1
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.5
multiprocess==0.70.16
mypy==1.10.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.18.0
openai==1.34.0
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.1
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.1.3
pandas-bokeh==0.5.5
panel==1.4.1
param==2.1.0
parso==0.8.3
patsy==0.5.6
pillow==10.3.0
pinecone-client==2.2.1
platformdirs==4.2.0
plotly==5.21.0
plotnine==0.13.4
pluggy==1.5.0
posthog==3.5.0
prompt_toolkit==3.0.45
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.16
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pygwalker==0.4.8
PyJWT==2.6.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproj==3.6.1
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyshp==2.3.1
PySocks==1.7.1
pyspark==3.5.1
pytest==8.2.0
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2022.7.1
pyviz_comms==3.0.2
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
questionary==2.0.1
referencing==0.33.0
regex==2023.10.3
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
segment-analytics-python==2.2.3
sentence-transformers==3.0.1
shapely==2.0.4
shellingham==1.5.4
shiny==0.9.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake-connector-python==3.9.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.5
SQLAlchemy==1.4.52
sqlglot==23.11.2
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.2
streamlit==1.32.0
streamlit-elements==0.1.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.12
sympy==1.12
tabpy==2.9.0
tbb==2021.12.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
textblob==0.18.0.post0
tf_keras==2.16.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.5.22
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.41.2
twilio==7.16.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.3
ujson==5.8.0
ultralytics==8.2.25
urllib3==1.26.14
utility==1.0
uvicorn==0.24.0.post1
virtualenv==20.26.2
wasmtime==12.0.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wikipedia==1.4.0
win32-setctime==1.1.0
wrapt==1.16.0
WTForms==3.1.2
xgboost==2.0.3
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.18.1
zope.interface==6.3
